Rename theme to neon
